+++ driver/src/main/java/oracle/nosql/driver/kv/StoreAccessTokenProvider.java
@@ -157,6 +157,10 @@ public class StoreAccessTokenProvider implements AuthorizationProvider {
      */
     private SslContext sslContext;
 
+    /*
+     * SSL handshake timeout in milliseconds;
+     */
+    private int sslHandshakeTimeoutMs;
     /**
      * @hidden
      * This is only used for unit test
@@ -392,6 +396,11 @@ public class StoreAccessTokenProvider implements AuthorizationProvider {
         return this;
     }
 
+    public StoreAccessTokenProvider setSslHandshakeTimeout(int timeout) {
+        this.sslHandshakeTimeoutMs = timeout;
+        return this;
+    }
+
     /**
      * Returns whether the provider is accessing a secured store
      *
@@ -474,7 +483,7 @@ public class StoreAccessTokenProvider implements AuthorizationProvider {
             client = new HttpClient(
                 loginHost, loginPort, 0, 0, 0,
                 (isSecure && !disableSSLHook) ? sslContext : null,
-                serviceName, logger);
+                 sslHandshakeTimeoutMs, serviceName, logger);
             return HttpRequestUtil.doGetRequest(
                 client,
                 NoSQLHandleConfig.createURL(endpoint, basePath + serviceName)